Change movie description

+++ MovieSelection.Models/Entities/Movie.cs
@@ -8,7 +8,7 @@ public class Movie
 
     public string? Description { get; set; }
 
-    public short Year { get; set; }
+    public int Year { get; set; }
 
     public int CountryId { get; set; }
