+++ MovieSelection.Models/RequestModels/PostMovie.cs
@@ -8,7 +8,7 @@ namespace MovieSelection.Models.RequestModels
 
         public string? Description { get; set; }
 
-        public short Year { get; set; }
+        public int Year { get; set; }
 
         public int CountryId { get; set; }
